requirements.txt
anyio==3.6.2
azure-core==1.26.4
azure-storage-blob==12.16.0
blinker==1.6.2
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
cryptography==40.0.2
fastapi==0.95.2
Flask==2.3.2
h11==0.14.0
httpcore==0.17.1
httpx==0.24.1
idna==3.4
importlib-metadata==6.6.0
isodate==0.6.1
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
pycparser==2.21
pydantic==1.10.7
requests==2.30.0
six==1.16.0
sniffio==1.3.0
starlette==0.27.0
typing-extensions==4.5.0
urllib3==2.0.2
uvicorn==0.22.0
Werkzeug==2.3.4
zipp==3.15.0
